grammar glowup in README.md
I tried to touch up the grammar and punctuation of the README file as much as I could.  

If you have any questions about any of my suggestions, please reach out!  I'm happy to try to explain, or offer alternatives if you don't find my suggestions to be ideal.  🥂 Cheers!